added constants. fixed minor bugs and added tests

+++ app/Http/Controllers/Api/V1/ReservationController.php
@@ -72,8 +72,13 @@ class ReservationController extends Controller
         $this->reservationService->updateReservation($request, $reservation);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Reservation $reservation): void
     {
+
+        $this->authorize('delete', Reservation::class);
         $this->reservationService->deleteReservation($reservation);
     }
 }
